Moved categories from Home to CategoriesList component
- Categories are now fetched from BE both on mobile and desktop versions

- Tests must be fixed
