+++ front/src/pages/Home.tsx
@@ -13,73 +13,6 @@ import {
   ResourcesList,
 } from '../components/organisms'
 import { Icon, Text, Title } from '../components/atoms'
-import { urls } from '../constants/urls'
-import axios from 'axios'
-import { useQuery } from '@tanstack/react-query'
-
-type Tcategories = {
-  id: number
-  category: string
-  resources: number
-  topics: number
-  img: string
-}
-
-const categories: Tcategories[] = [
-  {
-    id: 1,
-    img: icons.angular,
-    resources: 49,
-    category: 'Angular',
-    topics: 6,
-  },
-  {
-    id: 2,
-    img: icons.react,
-    category: 'React',
-    resources: 65,
-    topics: 7,
-  },
-  {
-    id: 3,
-    img: icons.vue,
-    category: 'Vue',
-    resources: 32,
-    topics: 8,
-  },
-  {
-    id: 4,
-    img: icons.javascript,
-    category: 'Javascript',
-    resources: 44,
-    topics: 3,
-  },
-  {
-    id: 5,
-    img: icons.dataScience,
-    category: 'Data Science',
-    resources: 23,
-    topics: 1,
-  },
-]
-
-// Categories from BE
-type TcategoriesBE = {
-  id: number
-  name: string
-}
-const fetchCategories = () => {
-  return axios.get(urls.categoriesURL)
-}
-
-const categoriesIcons: Record<string, string> = {
-  React: icons.react,
-  Angular: icons.angular,
-  Node: icons.node,
-  PHP: icons.php,
-  Java: icons.java,
-  'Data Science': icons.dataScience,
-}
 
 type Tresource = {
   id: number
@@ -142,13 +75,13 @@ const dataSubjects = [
 ]
 
 // TODO: mobile first!
-const MobileStyled = styled.div`
+export const MobileStyled = styled.div`
   display: block;
   @media only ${device.Laptop} {
     display: none;
   }
 `
-const DesktopStyled = styled.div`
+export const DesktopStyled = styled.div`
   display: none;
   @media only ${device.Laptop} {
     display: block;
@@ -189,15 +122,6 @@ const DivStyled = styled.div`
   width: 100%;
   height: 100%;
   border-radius: ${dimensions.borderRadius.base};
-`
-
-const CategoriesContainerStyled = styled(FlexBox)`
-  padding-left: ${dimensions.spacing.xxs};
-  padding-right: ${dimensions.spacing.xs};
-  margin-right: ${dimensions.spacing.sm};
-  align-items: flex-start;
-  color: ${colors.gray.gray3};
-  min-width: 11.5rem;
 `
 
 const UserResourcesContainerStyled = styled(FlexBox)`
@@ -246,7 +170,7 @@ const VideoContainer = styled(FlexBox)`
   padding-left: ${dimensions.spacing.xxxs};
 `
 
-const ImgStyled = styled.img`
+export const ImgStyled = styled.img`
   height: 30px;
   margin-right: ${dimensions.spacing.xxxs};
   margin-top: ${dimensions.spacing.xxl};
@@ -266,41 +190,15 @@ const LinkStyled = styled.a<TLinkStyled>`
   margin-top: ${dimensions.spacing.base};
   cursor: pointer;
 `
-const CategoryLinkStyled = styled.a<TLinkStyled>`
-  color: ${({ active }) => (active ? colors.black.black3 : colors.gray.gray3)};
-  font-weight: bold;
-  margin-top: ${dimensions.spacing.xxl};
-  cursor: pointer;
 
-  &::before {
-    content: '${({ active }) => (active ? '●' : '')}';
-    font-size: larger;
-    color: ${colors.primary};
-    margin-right: 0.3rem;
-  }
-`
 // END style Desktop
 
 const Home: FC = () => {
   const [activeLink, setActiveLink] = useState('')
-  const [activeCategory, setActiveCategory] = useState('')
+
   const handleClick = (link: SetStateAction<string>) => {
     setActiveLink(link)
   }
-  const handleCategoryClick = (cat: SetStateAction<string>) => {
-    setActiveCategory(cat)
-  }
-
-  const {
-    isLoading,
-    data: categoriesBE,
-    isError,
-  } = useQuery({
-    queryKey: ['categories'],
-    queryFn: fetchCategories,
-  })
-
-  console.log(categoriesBE?.data)
 
   return (
     <>
@@ -323,22 +221,7 @@ const Home: FC = () => {
       </MobileStyled>
       <DesktopStyled>
         <MainContainer>
-          <CategoriesContainerStyled>
-            {categoriesBE?.data.map((category: TcategoriesBE) => (
-              <FlexBox direction="row" key={category.id}>
-                <ImgStyled
-                  src={categoriesIcons[category.name]}
-                  alt={`${category.name} logo`}
-                />
-                <CategoryLinkStyled
-                  active={activeCategory === category.name}
-                  onClick={() => handleCategoryClick(category.name)}
-                >
-                  {category.name}
-                </CategoryLinkStyled>
-              </FlexBox>
-            ))}
-          </CategoriesContainerStyled>
+          <CategoriesList />
           {/* ==> CONTAINER CON LAS LAS COLUMNAS */}
           <DivStyled>
             {/* ==> COLUMNA BÚSQUEDA */}
